+++ components/pages/home/Unauthorized/CardSwiper.module.scss
@@ -1,0 +1,29 @@
+.swiper1 {
+    padding-left: 20px;
+    padding-right: 20px;
+    width: 100%;
+}
+
+.swiperSlide1 {
+    display: block;
+    object-fit: cover;
+    width: fit-content;
+    width: fit-content;
+    height: fit-content;
+    height: 164;
+}
+
+.swiper {
+    padding-left: 20px;
+    padding-right: 20px;
+    width: 100%;
+}
+
+.swiperSlide {
+    display: block;
+    object-fit: cover;
+    width: fit-content;
+    width: fit-content;
+    height: fit-content;
+    height: 400px;
+}
